Changed the way UTF8ToSHA1 was initialized
Because sha1 was initially a static instance variable, everytime a method that was called that was reliant on sha1, would call the first instance of sha1 which used old data.

By initializing sha1 in each method to call for prefix or suffix of the hash, sha1 would have "fresh" data to work with, passing all the unit tests.